+++ packages/frontend-sdk/src/sdk.ts
@@ -12,6 +12,8 @@ import {
   DepositQuoteResponse,
   DepositStatusResponse,
   JoinSplitProofWithPublicSignals,
+  OpDigestWithMetadata,
+  OperationMetadata,
   OperationRequest,
   OperationRequestBuilder,
   ProvenOperation,
@@ -367,6 +369,23 @@ export class NocturneFrontendSDK {
   }
 
   /**
+   * Return list of all inflight operation digests and metadata about each operation.
+   */
+  async getInflightOpDigestsWithMetadata(): Promise<OpDigestWithMetadata[]> {
+    const json = (await window.ethereum.request({
+      method: "wallet_invokeSnap",
+      params: {
+        snapId: SNAP_ID,
+        request: {
+          method: "nocturne_getInflightOpDigestsWithMetadata",
+        },
+      },
+    })) as string;
+
+    return JSON.parse(json) as OpDigestWithMetadata[];
+  }
+
+  /**
    * Invoke snap `syncNotes` method.
    */
   async sync(): Promise<void> {
@@ -396,7 +415,8 @@ export class NocturneFrontendSDK {
    * @param operationRequest Operation request
    */
   protected async requestSignOperation(
-    operationRequest: OperationRequest
+    operationRequest: OperationRequest,
+    opMetadata?: OperationMetadata
   ): Promise<SignedOperation> {
     const json = (await window.ethereum.request({
       method: "wallet_invokeSnap",
@@ -404,7 +424,10 @@ export class NocturneFrontendSDK {
         snapId: SNAP_ID,
         request: {
           method: "nocturne_signOperation",
-          params: { operationRequest: JSON.stringify(operationRequest) },
+          params: {
+            operationRequest: JSON.stringify(operationRequest),
+            opMetadata: opMetadata ? JSON.stringify(opMetadata) : undefined,
+          },
         },
       },
     })) as string;