+++ packages/sdk/src/index.ts
@@ -6,7 +6,7 @@ export * from "./crypto";
 export * from "./store";
 export * from "./sync";
 export * from "./request";
-export * from "./NullifierChecker";
+export * from "./OpTracker";
 
 export {
   OperationRequest,