+++ packages/sdk/test/prepareOperation.test.ts
@@ -125,9 +125,14 @@ describe("gatherNotes", () => {
     );
 
     // add optimistic NF records for the 30 token note at merkleIndex 0
-    await nocturneDB.storeOptimisticNFRecords(
-      [0],
-      [{ nullifier: 420n, expirationDate: Date.now() + 1_000_000 }]
+    await nocturneDB.storeOptimisticRecords(
+      0n,
+      {
+        expirationDate: Date.now() + 1_000_000,
+        merkleIndices: [0],
+        metadata: { description: "" },
+      },
+      [{ nullifier: 420n }]
     );
 
     // gather notes to spend 30 tokens total